+++ lib/convert_unit/convert_unit.rb
@@ -1,7 +1,7 @@
 module ConvertUnit
   def self.all_classes
-    self.constants.select do |k|
-      self.const_get(k).instance_of? Class
+    constants.select do |k|
+      const_get(k).instance_of? Class
     end
   end
 end