+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## v0.2.0
+* Change module required location
+* Add module self method for get available unit conversion class
 * Add Density method conversion
+* Add Volumetric liquid flow
 
 ## v0.1.0
 * initial release